requirements.txt
decorator>=4.1.0
imageio==2.4.1
matplotlib==3.0.1
numpy==1.15.3
networkx==2.1
pandas==0.23.4
plotly==3.4.0
pypng==0.0.18
pyqt5==5.11.3
requests==2.20.0
requests_futures==0.9.8
Shapely==1.6.4.post2
seaborn==0.9.0
setuptools==40.5.0
scipy==1.1.0
six==1.11.0
tqdm==4.28.1
vispy==0.5.3

# Below are optional dependencies
# py2cytoscape==0.7.0
# pyoctree==0.2.10
# ete3==3.0.0b35
# trimesh==2.35.2
# rpy2==2.9.4
# fuzzywuzzy[speedup]==0.17.0
